Add new data

Add new data

Usage

add_new_cells(
  new_raw_data_csv,
  cell_characteristics_csv,
  old_raw_data_csv,
  current_type,
  new_file_name
)

Arguments

new_raw_data_csv

A filepath to a csv containing the new raw data. If the data are evoked current data (current_type == "eEPSC") then this must contain 4 columns: Letter, ID, P1 and P2. If the data are spontaneous current data, the columns must be Letter, ID, Recording_num, Trace, Peak_amplitude and Time_of_peak. Please see the section below on required columns for more details.

cell_characteristics_csv

A filepath to a csv containing information about the cells. Please see import_cell_characteristics_df() for a description of what columns should be included.

old_raw_data_csv

A filepath to a csv containing the current raw data. Since this function appends the new data to the old data, this must be of the same current_type as the new data (e.g. the columns must be the same).

current_type

A character describing the current type. Allowed values are "eEPSC" or "sEPSC".

new_file_name

A filename for the csv containing the new data appended to the old data. Must be a character representing a filepath to a csv file. Examples include "Data/20241118-Raw-eEPSC-data.csv".

Value

A dataframe consisting of the old raw data with information from the new cells appended to it.

Examples

if (FALSE) add_new_cells(new_raw_data_csv = import_ext_data("sample_new_eEPSC_data.csv"),
cell_characteristics_csv = import_ext_data("sample_cell_characteristics.csv"),
old_raw_data_csv = import_ext_data("sample_eEPSC_data.csv"),
current_type = "eEPSC",
new_file_name = "20241118-Raw-eEPSC-Data.csv") # \dontrun{}