Summarize current data per 5-min for statistical tests

make_summary_EPSC_data() allows you to divide data from a long recording (e.g. 30 minutes) into evenly-spaced intervals (e.g. 5 minutes). It will generate summary data like the mean current amplitude for each interval. This can be useful for inserting into statistical models to compare effect sizes across broad stretches of time. The interval length would have been previously specified in make_normalized_EPSC_data() using the interval_length argument.

Usage

make_summary_EPSC_data(
  data = patchclampplotteR::sample_raw_eEPSC_df,
  current_type = "eEPSC",
  save_output_as_RDS = "no"
)

Arguments

data

A data.frame object generated using make_normalized_EPSC_data(). It must contain the columns outlined in the Required columns section below, which will already be generated for you from the output of make_normalized_EPSC_data().

current_type

A character describing the current type. Allowed values are "eEPSC" or "sEPSC".

save_output_as_RDS

A character ("yes" or "no") describing if the resulting object should be saved as an RDS file in the raw data folder.

Value

A dataframe with summary data such as the mean current amplitude, coefficient of variation, standard deviation, standard error, variance, variance-to-mean ratio, and inverse coefficient of variation squared for each interval.

letter A character value that is a unique identifier for a single recording. Used to link data sets for evoked or spontaneous currents and cell-characteristics.

synapses A character value (e.g. "Glutamate" or "GABA").

sex A character value (e.g. "Male" or "Female").

treatment A character value (e.g. "Control", "HNMPA").

time A numeric value that represents time in minutes. This column is autogenerated in add_new_cells()].

ID A character value for the recording filename.

P1 A numeric value representing the amplitude of the first evoked current in pA.

P2 A numeric value representing the amplitude of the second evoked current in pA.

X A numeric value representing the x-value of the cell's location in µm. Leave this blank if you don't have this data.

Y A numeric value representing the y-value of the cell's location in µm. Leave this blank if you don't have this data.

age A numeric value representing the animal's age. Can be any value as long as the time units are consistent throughout (e.g. don't mix up days and months when reporting animal ages).

animal A numeric value representing the animal's ID or number.

category A numeric value representing the experiment type. Used to assign top-level groups for further analyses, with treatment as subgroups.

cell A character or numeric value representing the cell. For example, use 3.1.1 for animal #3, slice #1, cell #1.

PPR (for evoked currents only) A numeric value that represents the paired pulse ratio (PPR) of the evoked currents, generated using dplyr::mutate(PPR = P2/P1).

interval A character value indicating the interval that the data point belongs to. For example, interval will be "t0to5" for any data points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.

baseline_range A logical value required for the baseline transformation. It is set to TRUE when time is within the baseline period (e.g. Time <= 5) and FALSE at all other times.

baseline_mean A numeric value representing the mean evoked current amplitude during the baseline period. There is a different baseline_mean for each letter.

P1_transformed A numeric value representing the first evoked current amplitude (pA) normalized relative to the mean amplitude during the recording's baseline.

P2_transformed A numeric value representing the second evoked current amplitude (pA) normalized relative to the mean amplitude during the recording's baseline.

Required basic columns

It doesn't matter if the data are evoked or current type - these columns should be included in your data.

letter A character value that is a unique identifier for a single recording. Used to link data sets for evoked or spontaneous currents and cell-characteristics.

synapses A character value (e.g. "Glutamate" or "GABA").

sex A character value (e.g. "Male" or "Female").

treatment A character value (e.g. "Control", "HNMPA").

time A numeric value that represents time in minutes. This column is autogenerated in add_new_cells().

ID A character value for the recording filename.

X A numeric value representing the x-value of the cell's location in µm. Leave this blank if you don't have this data.

Y A numeric value representing the y-value of the cell's location in µm. Leave this blank if you don't have this data.

age A numeric value representing the animal's age. Can be any value as long as the time units are consistent throughout (e.g. don't mix up days and months when reporting animal ages).

animal A numeric value representing the animal's ID or number.

category A numeric value representing the experiment type. Used to assign top-level groups for further analyses, with treatment as subgroups.

cell A character or numeric value representing the cell. For example, use 3.1.1 for animal #3, slice #1, cell #1.

Evoked current data:

If the data are evoked currents (current_type == "eEPSC"), the data must contain the basic columns mentioned in Required basic columns plus these columns:

P1 A numeric value representing the amplitude of the first evoked current in pA.

P2 A numeric value representing the amplitude of the second evoked current in pA.

Spontaneous current data:

If the data are spontaneous currents (current_type == "sEPSC"), the data must contain the basic columns mentioned in Required basic columns plus these columns:

recording_num A numeric value representing the recording number. This was incorporated before we switched to concatenating all recordings into one, but it needs to remain here to prevent breaking previous projects. It should be set to 1.

trace A numeric value representing the trace (automatically generated in Clampfit) where the current occured.

time_of_peak A numeric value representing the time of the peak in milliseconds relative to trace number. This is automatically calculated in Clampfit.

time A numeric value representing the absolute time when the current happened, relative to the start of the recording. This is autogenerated. See add_new_cells() for a description of how the true time value (time) is calculated from the recording_num and trace.

amplitude A numeric value representing the amplitude of the evoked current in pA.

Required evoked currents columns

These columns will all be generated automatically in make_normalized_EPSC_data(), but for more details, you can look at sample_raw_eEPSC_df to see an example of what the incoming raw dataset in the data argument should look like. If the data are evoked currents (current_type == "eEPSC"), the data must contain the basic columns mentioned in Required basic columns plus these columns:

PPR A numeric value that represents the paired pulse ratio (PPR) of the evoked currents, generated in make_normalized_EPSC_data().

interval A character value indicating the interval that the data belong to (e.g. "t0to5" for the first 5 minutes, "t5to10"). Generated automatically in make_normalized_EPSC_data().

baseline_range A logical value required for the baseline transformation. It is set to TRUE when time is within the baseline period (e.g. time <= 5) and FALSE at all other times. Generated automatically in make_normalized_EPSC_data().

baseline_mean A numeric value representing the mean evoked current amplitude during the baseline period. There is a different baseline_mean for each letter. Generated automatically in make_normalized_EPSC_data().

P1_transformed A numeric value representing the first evoked current amplitude (pA) normalized relative to the mean amplitude during the recording's baseline. Generated automatically in make_normalized_EPSC_data().

P2_transformed A numeric value representing the second evoked current amplitude (pA) normalized relative to the mean amplitude during the recording's baseline. Generated automatically in make_normalized_EPSC_data().

Required spontaneous currents columns

If the data are spontaneous currents (current_type == "sEPSC"), the data must contain the basic columns mentioned in Required basic columns plus these columns:

interval A character value indicating the interval that the data belong to (e.g. "t0to5" for the first 5 minutes, "t5to10"). Generated automatically in make_normalized_EPSC_data().

baseline_range A logical value required for the baseline transformation. It is set to TRUE when time is within the baseline period (e.g. Time <= 5) and FALSE at all other times. Generated automatically in make_normalized_EPSC_data().

baseline_mean A numeric value representing the mean evoked current amplitude during the baseline period. There is a different baseline_mean for each letter. Generated automatically in make_normalized_EPSC_data().

amplitude_transformed A numeric value representing the spontaneous current amplitude (pA) normalized relative to the mean amplitude during the recording's baseline. Generated automatically in make_normalized_EPSC_data().

Examples

make_summary_EPSC_data(
  data = sample_raw_eEPSC_df,
  current_type = "eEPSC"
)
#> # A tibble: 95 × 20
#>    category letter sex    treatment interval mean_P1_transformed mean_P1_raw
#>    <fct>    <fct>  <fct>  <fct>     <fct>                  <dbl>       <dbl>
#>  1 2        AO     Male   Control   t0to5                  100         36.8 
#>  2 2        AO     Male   Control   t5to10                  34.0       12.5 
#>  3 2        AO     Male   Control   t10to15                 17.7        6.49
#>  4 2        AO     Male   Control   t15to20                 18.6        6.85
#>  5 2        AO     Male   Control   t20to25                 21.3        7.82
#>  6 2        AZ     Female Control   t0to5                  100         44.3 
#>  7 2        AZ     Female Control   t5to10                  74.1       32.8 
#>  8 2        AZ     Female Control   t10to15                 53.7       23.8 
#>  9 2        AZ     Female Control   t15to20                 56.0       24.8 
#> 10 2        AZ     Female Control   t20to25                 50.7       22.4 
#> # ℹ 85 more rows
#> # ℹ 13 more variables: n <int>, sd <dbl>, cv <dbl>, se <dbl>,
#> #   cv_inverse_square <dbl>, variance <dbl>, VMR <dbl>, age <int>,
#> #   animal <dbl>, X <dbl>, Y <dbl>, time <dbl>, synapses <fct>