Representative data from a series of evoked (eEPSC) currents

This is an example of the raw eEPSC data produced using make_normalized_EPSC_data(). It is useful for demonstrating functions that build off of this dataset, such as make_pruned_EPSC_data() and plotting functions like plot_raw_current_data() and plot_summary_current_data().

Format

An .rda file containing 5680 objects of 20 variables.

letter A character value that is a unique identifier for a single recording. Used to link data sets for evoked or spontaneous currents and cell-characteristics. synapses A character value (e.g. "Glutamate" or "GABA"). sex A character value (e.g. "Male" or "Female"). treatment A character value (e.g. "Control", "HNMPA") representing the antagonists or agonists applied, or any protocol applied to the animals (e.g. "Fasting"). time A numeric value that represents time in minutes. This column is autogenerated in add_new_cells()]. ID A character value for the recording filename. P1 A numeric value representing the amplitude of the first evoked current in pA. P2 A numeric value representing the amplitude of the second evoked current in pA. X A numeric value representing the x-value of the cell's location in µm. Leave this blank if you don't have this data. Y A numeric value representing the y-value of the cell's location in µm. Leave this blank if you don't have this data. age A numeric value representing the animal's age. Can be any value as long as the time units are consistent throughout (e.g. don't mix up days and months when reporting animal ages). animal A numeric value representing the animal's ID or number. category A numeric value representing the experiment type. Used to assign top-level groups for further analyses, with treatment as subgroups. cell A character or numeric value representing the cell. For example, use 3.1.1 for animal #3, slice #1, cell #1. PPR (for evoked currents only) A numeric value that represents the paired pulse ratio (PPR) of the evoked currents, generated using dplyr::mutate(PPR = P2/P1). interval A character value indicating the interval that the data point belongs to. For example, interval will be "t0to5" for any data points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc. baseline_range A logical value required for the baseline transformation. It is set to TRUE when time is within the baseline period (e.g. Time <= 5) and FALSE at all other times. baseline_mean A numeric value representing the mean evoked current amplitude during the baseline period. There is a different baseline_mean for each letter. P1_transformed A numeric value representing the first evoked current amplitude (pA) normalized relative to the mean amplitude during the recording's baseline. P2_transformed A numeric value representing the second evoked current amplitude (pA) normalized relative to the mean amplitude during the recording's baseline.

See also