Representative data from a series of spontaneous (sEPSC) currents

This is an example of the raw spontaneous current data produced using make_normalized_EPSC_data(). It is useful for demonstrating functions that build off of this dataset, such as make_pruned_EPSC_data() and plotting functions like plot_raw_current_data() and plot_summary_current_data().

Format

An .rda file containing 5680 objects of 20 variables.

letter

A character value that is a unique identifier for a single recording. Used to link data sets for evoked or spontaneous currents and cell-characteristics.

ID

A character value for the recording filename.

recording_num

A numeric value representing the recording number. This was incorporated before we switched to concatenating all recordings into one, but it needs to remain here to prevent breaking previous projects. It should be set to 1.

trace

A numeric value representing the trace (automatically generated in Clampfit) where the current occurred.

time_of_peak

A numeric value representing the time of the peak in milliseconds relative to trace number. This is automatically calculated in Clampfit.

amplitude

A numeric value representing the amplitude of the evoked current in pA.

cell

A character or numeric value representing the cell. For example, use 3.1.1 for animal #3, slice #1, cell #1.

sex

A character value (e.g. "Male" or "Female").

X

A numeric value representing the x-value of the cell's location in µm. Leave this blank if you don't have this data.

Y

A numeric value representing the y-value of the cell's location in µm. Leave this blank if you don't have this data.

age

A numeric value representing the animal's age. Can be any value as long as the time units are consistent throughout (e.g. don't mix up days and months when reporting animal ages).

animal

A numeric value representing the animal's ID or number.

synapses

A character value (e.g. "Glutamate" or "GABA").

treatment

A character value (e.g. "Control", "HNMPA") representing the antagonists or agonists applied, or any protocol applied to the animals (e.g. "Fasting").

category

A numeric value representing the experiment type. Used to assign top-level groups for further analyses, with treatment as subgroups.

time

A numeric value representing the absolute time when the current happened, relative to the start of the recording. This is autogenerated. See [add_new_cells() for a description of how the true time value (time) is calculated from the recording_num and trace.

interval

A character value indicating the interval that the data point belongs to. For example, interval will be "t0to5" for any data points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.

baseline_range

A logical value required for the baseline transformation. It is set to TRUE when time is within the baseline period (e.g. Time <= 5) and FALSE at all other times.

baseline_mean

A numeric value representing the mean evoked current amplitude during the baseline period. There is a different baseline_mean for each letter.

amplitude_transformed

A numeric value representing the spontaneous current amplitude (pA) normalized relative to the mean spontaneous current amplitude during the recording's baseline.

See also