Summarize current data per 5-min for statistical tests

make_summary_EPSC_data() allows you to divide data from a long recording (e.g. 30 minutes) into evenly-spaced intervals (e.g. 5 minutes). It will generate summary data like the mean current amplitude for each interval. This can be useful for inserting into statistical models to compare effect sizes across broad stretches of time. The interval length would have been previously specified in make_normalized_EPSC_data() using the interval_length argument.

Usage

make_summary_EPSC_data(
  data = patchclampplotteR::sample_raw_eEPSC_df,
  current_type = "eEPSC",
  save_output_as_RDS = "no",
  decimal_places = 2,
  baseline_interval = "t0to5",
  ending_interval = "t20to25"
)

Arguments

data

A data.frame object. If the data are evoked currents (current_type == "eEPSC"), this should be the raw evoked current data generated using make_normalized_EPSC_data(). If the data are spontaneous currents (current_type == "sEPSC"), this should be the pruned data $individual_cells dataset generated using make_pruned_EPSC_data().

current_type

A character describing the current type. Allowed values are "eEPSC" or "sEPSC".

save_output_as_RDS

A character ("yes" or "no") describing if the resulting object should be saved as an RDS file in the raw data folder.

decimal_places

A numeric value indicating the number of decimal places the data should be rounded to. Used to reduce file size and prevent an incorrect representation of the number of significant digits.

baseline_interval

A character value describing the baseline interval. Defaults to "t0to5".

ending_interval

A character value describing the last interval in the recording. Useful for future plots in which you compare the percent decrease/increase in current amplitude relative to the baseline. Examples include "t20to25", "t10to15", etc.

Value

A dataframe with summary data such as the mean current amplitude, coefficient of variation, standard deviation, standard error, variance, variance-to-mean ratio, and inverse coefficient of variation squared for each interval.

New columns for evoked current data (current_type == "eEPSC") include:

mean_P1_transformed The amplitude of the first evoked current amplitude (% Baseline eEPSC amplitude) normalized to the mean baseline amplitude and averaged over the interval.

mean_P1_raw The amplitude of the first evoked current amplitude (pA) averaged over the interval.

n The number of datapoints used to create the averaged values. Corresponds to the number of sweeps per interval.

sd The standard deviation of the normalized evoked current data (P1_transformed).

cv The coefficient of variation of P1_transformed.

se The standard error of P1_transformed.

cv_inverse_square The inverse of the squared coefficient of variation of P1_transformed.

variance The variance of P1_transformed.

VMR The variance-to-mean ratio (VMR) of P1_transformed.

interval A character value indicating the interval that the data point belongs to. For example, interval will be "t0to5" for any data points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.

letter, synapses, sex, treatment, etc. Unmodified columns from the original dataset describing the cell's properties.

New columns for spontaneous current data (current_type == "sEPSC") include:

mean_transformed_amplitude The average normalized spontaneous current amplitude (% Baseline sEPSC amplitude).

mean_raw_amplitude The average raw spontaneous current amplitude (pA).

n The number of datapoints used to create the average.

sd_transformed_amplitude The standard deviation of the normalized spontaneous current data (mean_transformed_amplitude).

se_transformed_amplitude The standard error of mean_transformed_amplitude.

mean_transformed_frequency The average normalized frequency (% Baseline frequency).

sd_transformed_frequency The standard deviation of mean_transformed_frequency.

se_frequency The standard error of mean_transformed_frequency.

mean_raw_frequency The average raw frequency (Hz).

letter, synapses, sex, treatment, etc. Unmodified columns from the original dataset describing the cell's properties.

See also

Examples

make_summary_EPSC_data(
  data = sample_raw_eEPSC_df,
  current_type = "eEPSC",
  save_output_as_RDS = "no",
  decimal_places = 2
)
#> $percent_change_data
#> # A tibble: 19 × 15
#> # Groups:   category, treatment, letter [19]
#>    category letter sex    treatment       age animal     X     Y synapses  t0to5
#>    <fct>    <fct>  <fct>  <fct>         <dbl>  <dbl> <dbl> <dbl> <fct>     <dbl>
#>  1 2        AO     Male   Control          39   17     NA    NA  Glutamate  36.8
#>  2 2        AZ     Female Control          32   21    353.  332. Glutamate  44.3
#>  3 2        BN     Male   Control          29   25    153.  337. Glutamate  72.1
#>  4 2        BO     Male   HNMPA            32   27     NA    NA  Glutamate  87.7
#>  5 2        BT     Male   HNMPA            37   30    387.  587. Glutamate  51.3
#>  6 2        CG     Female HNMPA            36   35    164.  367. Glutamate  44.4
#>  7 2        CZ     Male   HNMPA            28   41    297.  493. Glutamate  68.1
#>  8 2        FT     Female HNMPA            39   72    153.  337. Glutamate  77.8
#>  9 2        FX     Female PPP              28   74    235.  496. Glutamate  61.3
#> 10 2        GF     Female PPP              36   77    217.  324. Glutamate  79.2
#> 11 2        GI     Female PPP              28   81    236.  284. Glutamate  41.0
#> 12 2        GK     Male   PPP              35   84    249.  574. Glutamate  69.1
#> 13 2        GR     Male   PPP              34   90    314.  416. Glutamate  99.1
#> 14 2        GX     Male   PPP_and_HNMPA    33   97     NA    NA  Glutamate  49.9
#> 15 2        HB     Male   PPP_and_HNMPA    39  100    133.  590. Glutamate  49.4
#> 16 2        HC     Male   PPP_and_HNMPA    39  100    173.  576. Glutamate  31.3
#> 17 2        HG     Female PPP_and_HNMPA    34  103     NA    NA  Glutamate  79.9
#> 18 2        HN     Male   PPP_and_HNMPA    38  109    289.  400. Glutamate  86.6
#> 19 2        L      Male   Control          38    8.5   NA    NA  Glutamate  77.0
#> # ℹ 5 more variables: t5to10 <dbl>, t10to15 <dbl>, t15to20 <dbl>,
#> #   t20to25 <dbl>, percent_change <dbl>
#> 
#> $summary_data
#> # A tibble: 95 × 20
#>    category letter sex    treatment interval mean_P1_transformed mean_P1_raw
#>    <fct>    <fct>  <fct>  <fct>     <fct>                  <dbl>       <dbl>
#>  1 2        AO     Male   Control   t0to5                  100         36.8 
#>  2 2        AO     Male   Control   t5to10                  34.0       12.5 
#>  3 2        AO     Male   Control   t10to15                 17.7        6.49
#>  4 2        AO     Male   Control   t15to20                 18.6        6.85
#>  5 2        AO     Male   Control   t20to25                 21.3        7.82
#>  6 2        AZ     Female Control   t0to5                  100         44.3 
#>  7 2        AZ     Female Control   t5to10                  74.1       32.8 
#>  8 2        AZ     Female Control   t10to15                 53.7       23.8 
#>  9 2        AZ     Female Control   t15to20                 56.0       24.8 
#> 10 2        AZ     Female Control   t20to25                 50.7       22.4 
#> # ℹ 85 more rows
#> # ℹ 13 more variables: n <dbl>, sd <dbl>, cv <dbl>, se <dbl>,
#> #   cv_inverse_square <dbl>, variance <dbl>, VMR <dbl>, age <dbl>,
#> #   animal <dbl>, X <dbl>, Y <dbl>, time <dbl>, synapses <fct>
#> 

make_summary_EPSC_data(
  data = sample_pruned_sEPSC_df$individual_cells,
  current_type = "sEPSC",
  save_output_as_RDS = "no",
  decimal_places = 2
)
#> # A tibble: 35 × 16
#> # Groups:   category, letter, sex, treatment [7]
#>    category letter sex    treatment interval mean_transformed_amplitude
#>    <fct>    <fct>  <fct>  <fct>     <fct>                         <dbl>
#>  1 2        AZ     Female Control   t0to5                         100. 
#>  2 2        AZ     Female Control   t5to10                         96.6
#>  3 2        AZ     Female Control   t10to15                        94.5
#>  4 2        AZ     Female Control   t15to20                        94.6
#>  5 2        AZ     Female Control   t20to25                        88.3
#>  6 2        BO     Male   HNMPA     t0to5                         100. 
#>  7 2        BO     Male   HNMPA     t5to10                         89.3
#>  8 2        BO     Male   HNMPA     t10to15                        81.9
#>  9 2        BO     Male   HNMPA     t15to20                        82.2
#> 10 2        BO     Male   HNMPA     t20to25                        82.0
#> # ℹ 25 more rows
#> # ℹ 10 more variables: mean_raw_amplitude <dbl>,
#> #   sd_transformed_amplitude <dbl>, n <dbl>, se_transformed_amplitude <dbl>,
#> #   mean_transformed_frequency <dbl>, sd_transformed_frequency <dbl>,
#> #   se_transformed_frequency <dbl>, mean_raw_frequency <dbl>, time <dbl>,
#> #   synapses <fct>