A dataframe of theme options for things like colours and line widths

This is an example of the dataframe used to specify the colour themes and plot options (like line thickness).

Format

An .rda file containing 23 objects of 2 variables

gray_shading_colour Hex code for the colour used to fill the violin plots. line_col Hex code for the colour used for the line indicating periods when a hormone was applied. rectangle_shading_colour Hex code for the colour used to shade the rectangles. plot_font_family plot_light_font_family significance_stars_font male_shape Numeric value representing the point shape used for data values where Sex == Male. female_shape Numeric value representing the point shape used for data values where Sex == Female. mean_point_colour Hex code for the colour used to fill the violin plots baseline_group_colour post_hormone_group_colour connecting_line_width connecting_line_colour_aps connecting_line_width_PPR mean_point_size geom_sina_size geom_signif_text_size AP_trace_size scale_bar_shift_y scale_bar_shift_x y_axis_limit "Control", "HNMPA"). line_col A character value with the same treatment names as in treatment except you would replace underscores with spaces and add line breaks if needed. This is to create more attractive, human-readable axis labels and table values.